examples/blob-and-file-triggers/blob_csv_to_table/requirements.txt
azure-functions==1.21.3
azure-data-tables==12.6.0
azure-functions-logging==0.5.0
azure-storage-blob==12.25.1
